pyproject.toml
[project]
name = "croustillantdatasets"
version = "1.0.0"
description = "CROUStillant Datasets"
authors = ["Paul Bayfield"]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "async-timeout>=5.0.1",
    "asyncpg>=0.30.0",
    "datagouvpy",
    "discord-py>=2.5.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
]

[tool.uv.sources]
datagouvpy = { git = "https://github.com/CROUStillant-Developpement/DataGouvPy" }
